FIR IDE: Move ReplaceCallFix to idea-frontend-independent.
